Move ChangePolicy to drain.rs
