+++ src/lib.rs
@@ -22,8 +22,6 @@ pub mod metrics;
 
 mod feerate;
 pub use feerate::*;
-mod change_policy;
-pub use change_policy::*;
 mod target;
 pub use target::*;
 mod drain;
